requirements.txt
astroid==2.3.3
astropy==4.0.1.post1
atomicwrites==1.3.0
attr==0.3.1
attrs==19.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
confuse==1.0.0
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
htmlmin==0.1.12
idna==2.8
importlib-metadata==1.6.0
ipywidgets==7.5.1
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
jupyter-client==6.1.2
kaggle==1.5.6
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
line-bot-sdk==1.16.0
llvmlite==0.31.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
missingno==0.4.2
more-itertools==8.2.0
numba==0.48.0
numpy==1.18.3
pandas==1.0.3
pandas-datareader==0.8.1
pandas-profiling==2.5.0
phik==0.9.9
pluggy==0.13.1
py==1.8.1
pylint==2.4.4
pytest==5.4.1
pytest-pylint==0.15.1
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==2.0
tangled-up-in-unicode==0.0.3
text-unidecode==1.3
tqdm==4.42.0
twstock==1.3.1
typed-ast==1.4.1
urllib3==1.25.8
visions==0.2.2
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0
